more organized code
