+++ ui/src/components/General/ListItems/News/news.types.ts
@@ -1,0 +1,7 @@
+interface NewsType {
+	title: string;
+	content: string;
+	url: string;
+}
+
+export type { NewsType };
