+++ ui/src/components/General/ListItems/Resource/resource.type.ts
@@ -1,0 +1,8 @@
+interface ResourceType {
+	name: string;
+	category: string;
+	subCategory: string[];
+	url: string;
+}
+
+export type { ResourceType };
