+++ ui/src/pages/IconList/index.tsx
@@ -3,6 +3,9 @@ import Resource from "components/General/ListItems/Resource/Resource";
 import { useQuery } from "@tanstack/react-query";
 
 import { getData } from "api/API";
+import { ListItemProps } from "components/Hoc/List/types";
+import { ResourceType } from "components/General/ListItems/Resource/resource.type";
+import { NewsType } from "components/General/ListItems/News/news.types";
 
 const URL = `./icons.json`;
 const QUERY_KEY = "icons";
@@ -18,7 +21,9 @@ function IconList() {
 		<List
 			items={data}
 			resourceName="github"
-			itemComponent={Resource}
+			itemComponent={
+				Resource as React.FC<ListItemProps<ResourceType | NewsType>>
+			}
 			isLoading={isLoading}
 			isError={isError}
 		/>
