+++ ui/src/pages/News/index.tsx
@@ -3,6 +3,9 @@ import { getData } from "api/API";
 import { APIResponse } from "./types";
 import List from "components/Hoc/List/List";
 import News from "components/General/ListItems/News/News";
+import { NewsType } from "components/General/ListItems/News/news.types";
+import { ResourceType } from "components/General/ListItems/Resource/resource.type";
+import { ListItemProps } from "components/Hoc/List/types";
 
 const URL = `./news.json`;
 const QUERY_KEY = "news";
@@ -17,9 +20,11 @@ const NewsPage = () => {
 
 	return (
 		<List
-			items={data?.articles}
+			items={data?.articles || []}
 			resourceName="news"
-			itemComponent={News}
+			itemComponent={
+				News as React.FC<ListItemProps<ResourceType | NewsType>>
+			}
 			isLoading={isLoading}
 			isError={isError}
 		/>
