+++ ui/src/pages/YouTubeList/index.tsx
@@ -2,6 +2,9 @@ import List from "components/Hoc/List/List";
 import Resource from "components/General/ListItems/Resource/Resource";
 import { useQuery } from "@tanstack/react-query";
 import { getData } from "api/API";
+import { NewsType } from "components/General/ListItems/News/news.types";
+import { ListItemProps } from "components/Hoc/List/types";
+import { ResourceType } from "components/General/ListItems/Resource/resource.type";
 
 const URL = `./youtube.json`;
 const QUERY_KEY = "youtube";
@@ -17,7 +20,9 @@ function YoutubeList() {
 		<List
 			items={data}
 			resourceName={QUERY_KEY}
-			itemComponent={Resource}
+			itemComponent={
+				Resource as React.FC<ListItemProps<ResourceType | NewsType>>
+			}
 			isLoading={isLoading}
 			isError={isError}
 		/>
